requirements.txt
crewai
tavily-python
python-dotenv
python-docx==1.1.2
PyPDF2
fpdf